More meaningful timeout logs

+++ chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/scenario/Calibrator.java
@@ -28,6 +28,6 @@ import java.util.Map;
 
 public interface Calibrator {
 
-    void calibrate(String sessionId, String taskId, ScenarioFactory<Object, Object, Object> scenarioFactory, Map<NodeId, RemoteExecutor> remotes, int timeout);
+    void calibrate(String sessionId, String taskId, ScenarioFactory<Object, Object, Object> scenarioFactory, Map<NodeId, RemoteExecutor> remotes, long timeout);
 
 }
